requirements.txt
# === Ana kütüphaneler ===
gradio==5.49.1
chromadb==1.2.1
sentence-transformers==5.1.1
huggingface-hub==0.35.3
tqdm==4.67.1
python-dotenv==1.1.1

# === Gemini API (her iki SDK varyantı da destekleniyor) ===
google-generativeai==0.8.5
google-genai==1.46.0

# === Yardımcı bağımlılıklar ===
protobuf==5.29.5
typing-extensions==4.15.0
requests==2.32.5